+++ R/cross_reference.r
@@ -2,9 +2,7 @@
 #'
 #' @param schema a database schema without relationships, as given by
 #'   \code{\link{synthesise}}.
-#' @param ensure_lossless a logical, TRUE by default. If TRUE, and the
-#'   decomposition isn't lossless, an extra relation is added to make the
-#'   decomposition lossless.
+#' @inheritParams normalise
 #'
 #' @return A database schema with relationships, represented by a named list of
 #'   three lists and two character vectors, with the first four having equal