+++ R/synthesise.r
@@ -18,11 +18,7 @@
 #'
 #' Relation names are adjusted for uniqueness after synthesis.
 #'
-#' @param dependencies a list of functional dependencies, as given by
-#'   \code{\link{discover}}: each dependency is a list, containing one character
-#'   vector for the determinant set, and one unit-length character vector for
-#'   the dependent.
-#' @inheritParams autodb
+#' @inheritParams normalise
 #'
 #' @return A database schema, represented by a named list of two lists and two
 #'   character vectors: