+++ tests/testthat/test-cross_reference.r
@@ -22,31 +22,30 @@ describe("cross_reference", {
     expect_identical(database$relationships, expected_relations)
   })
   it("gives valid schemas", {
+    # same as test for normalise, need synthesis result generator
     # table_dum and table_dee
     empty_fds <- functional_dependency(list(), attrs = character())
-    empty_schema <- cross_reference(synthesise(empty_fds))
+    empty_schema <- normalise(empty_fds)
     is_valid_database_schema(empty_schema)
 
     forall(
       gen_flat_deps(7, 20),
       apply_both(
-        synthesise %>>% cross_reference %>>% is_valid_database_schema,
-        synthesise %>>%
-          with_args(cross_reference, ensure_lossless = FALSE) %>>%
+        normalise %>>% is_valid_database_schema,
+        with_args(normalise, ensure_lossless = FALSE) %>>%
           is_valid_database_schema
       )
     )
   })
   it("only links children to parents by exactly one parent key", {
     links_by_exactly_one_parent_key <- function(deps) {
-      schema <- synthesise(deps)
+      schema <- normalise(deps)
       if (length(schema$keys) <= 1)
         discard()
-      linked <- cross_reference(schema)
-      if (length(linked$relationships) == 0)
+      if (length(schema$relationships) == 0)
         discard()
-      relationship_tables <- lapply(linked$relationships, `[[`, 1)
-      relationship_attrs <- vapply(linked$relationships, `[[`, character(1), 2)
+      relationship_tables <- lapply(schema$relationships, `[[`, 1)
+      relationship_attrs <- vapply(schema$relationships, `[[`, character(1), 2)
       tables_index <- as.data.frame(do.call(rbind, relationship_tables))
       link_sets <- tapply(
         relationship_attrs,
@@ -62,7 +61,7 @@ describe("cross_reference", {
           logical(1)
         )])
         expect_length(
-          setdiff(attribute_sets, lapply(linked$keys[[parent]], sort)),
+          setdiff(attribute_sets, lapply(schema$keys[[parent]], sort)),
           0
         )
       }
@@ -77,8 +76,7 @@ describe("cross_reference", {
     reintroduces_missing_attrs_if_lossless <- function(deps) {
       lone_attr <- LETTERS[length(attrs(deps)) + 1]
       attr(deps, "attrs") <- c(attrs(deps), lone_attr)
-      schema <- synthesise(deps)
-      linked <- cross_reference(schema, ensure_lossless = TRUE)
+      linked <- normalise(deps, ensure_lossless = TRUE)
       expect_true(lone_attr %in% unlist(linked$attrs))
     }
     forall(
@@ -88,30 +86,25 @@ describe("cross_reference", {
   })
   it("has no change in added table for losslessness if avoidable attributes removed", {
     still_lossless_with_less_or_same_attributes_dep <- function(flat_deps) {
-      norm_deps_avoid <- synthesise(
+      schema_avoid_lossy <- normalise(
         flat_deps,
+        ensure_lossless = FALSE,
         remove_avoidable = TRUE
       )
-      norm_deps_noavoid <- synthesise(
+      schema_noavoid_lossy <- normalise(
         flat_deps,
+        ensure_lossless = FALSE,
         remove_avoidable = FALSE
       )
-
-      schema_avoid_lossy <- cross_reference(
-        norm_deps_avoid,
-        ensure_lossless = FALSE
+      schema_avoid_lossless <- normalise(
+        flat_deps,
+        ensure_lossless = TRUE,
+        remove_avoidable = TRUE
       )
-      schema_noavoid_lossy <- cross_reference(
-        norm_deps_noavoid,
-        ensure_lossless = FALSE
-      )
-      schema_avoid_lossless <- cross_reference(
-        norm_deps_avoid,
-        ensure_lossless = TRUE
-      )
-      schema_noavoid_lossless <- cross_reference(
-        norm_deps_noavoid,
-        ensure_lossless = TRUE
+      schema_noavoid_lossless <- normalise(
+        flat_deps,
+        ensure_lossless = TRUE,
+        remove_avoidable = FALSE
       )
       lengths_avoid_lossy <- lengths(schema_avoid_lossy$attrs)
       lengths_noavoid_lossy <- lengths(schema_noavoid_lossy$attrs)
@@ -163,7 +156,7 @@ describe("cross_reference", {
   })
   it("adds table with key with attributes in original order", {
     adds_ordered_primary_keys <- function(fds) {
-      schema <- cross_reference(synthesise(fds), ensure_lossless = TRUE)
+      schema <- normalise(fds, ensure_lossless = TRUE)
       all_keys <- unlist(schema$keys, recursive = FALSE)
       key_indices <- lapply(all_keys, match, schema$all_attrs)
       expect_false(any(vapply(key_indices, is.unsorted, logical(1))))
@@ -175,8 +168,7 @@ describe("cross_reference", {
   })
   it("only returns non-extraneous table relationships", {
     only_returns_non_extraneous_relationships <- function(deps) {
-      schema <- synthesise(deps)
-      linked <- cross_reference(schema, ensure_lossless = TRUE)
+      linked <- normalise(deps, ensure_lossless = TRUE)
       table_relationships <- unique(lapply(linked$relationships, `[[`, 1))
       table_relationships <- list(
         determinant_sets = lapply(table_relationships, `[`, 1),
@@ -195,8 +187,7 @@ describe("cross_reference", {
   it("is idempotent", {
     forall(
       gen_flat_deps(7, 20),
-      synthesise %>>%
-        cross_reference %>>%
+      normalise %>>%
         expect_biidentical(identity, cross_reference)
     )
   })
@@ -233,7 +224,7 @@ describe("cross_reference", {
           ),
         relation$attrs[[1]]
       )
-      redo <- cross_reference(synthesise(deps), ensure_lossless = TRUE)
+      redo <- normalise(deps, ensure_lossless = TRUE)
       expect_length(redo$attrs, 1)
       expect_identical(redo$attrs[[1]], relation$attrs[[1]])
       expect_identical(redo$keys[[1]], relation$keys[[1]])